added a concluding paragraph to the Java section
When I as a first-time reader arrived at the end of that section, it
didn't feel like the section really ended here, rather as if the XML
section was a follow-up section. I was in the mood of delving down into
the details.

Therefore, I added the paragraph that encourages the reader to explore
for themselves what do visualize next.